Add more boundaries for histogram
This is so it can be used with both seconds and milliseconds.
Milliseconds ceil at 1s, seconds have a lot of room now (up to 1000s).

Until I have figured out how to provide custom histogram boundaries for
each recorder, we have to live with this verbose set for now.